removed data collection activity
