+++ src/app/app/src/main/AndroidManifest.xml
@@ -26,7 +26,6 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup,GoogleAppIndexingWarning">
-        <activity android:name=".DataCollectionActivity" />
         <activity android:name=".QuizActivity" />
         <activity
             android:name=".MainActivity"
