+++ src/app/app/src/main/java/com/anguel/dissertation/App.java
@@ -2,6 +2,7 @@ package com.anguel.dissertation;
 
 import android.app.Application;
 
+import com.anguel.dissertation.serviceengine.ServiceEngine;
 import com.jakewharton.threetenabp.AndroidThreeTen;
 
 public class App extends Application {
@@ -9,5 +10,6 @@ public class App extends Application {
     public void onCreate() {
         super.onCreate();
         AndroidThreeTen.init(this);
+        ServiceEngine.getInstance(this);
     }
 }
