--- src/app/app/src/main/res/layout/activity_data_collection.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".DataCollectionActivity">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginTop="8dp"
-        android:layout_marginBottom="8dp"
-        android:orientation="vertical"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <TextView
-            android:id="@+id/textView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/use_the_button_below_to_turn_on_and_off_the_data_collection"
-            android:textAlignment="center"
-            android:textColor="#000000"
-            android:textSize="18sp" />
-
-        <ToggleButton
-            android:id="@+id/toggleButton"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="8dp"
-            android:layout_marginRight="8dp"
-            android:checked="false"
-            android:contentDescription="@string/toggle_data_collection"
-            android:text="@string/toggle_data_collection"
-            android:textAlignment="center"
-            android:textOff="@string/data_collection_is_off"
-            android:textOn="@string/data_collection_is_on" />
-
-<!--        todo: this is where users can be shown the data collected-->
-
-    </LinearLayout>
-</androidx.constraintlayout.widget.ConstraintLayout>
