+++ src/app/app/src/main/res/layout/content_main.xml
@@ -43,13 +43,16 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/welcomeText" />
 
-    <Button
+    <ToggleButton
         android:id="@+id/startDataCollectionButton"
         android:layout_width="0dp"
         android:layout_height="60dp"
         android:layout_marginStart="8dp"
         android:layout_marginEnd="8dp"
-        android:text="@string/see_data_collection"
+        android:checked="false"
+        android:text="@string/data_collection_is_off"
+        android:textOff="@string/data_collection_is_off"
+        android:textOn="@string/data_collection_is_on"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
